refactor: group dashboard widget components
