+++ frontend/src/dashboard/DashboardPage.tsx
@@ -1,12 +1,12 @@
 import { useAuth } from '@/hooks/useAuth'
-import { CreditWidget } from '@/dashboard/widgets/CreditWidget'
-import { NetWorthWidget } from '@/dashboard/widgets/NetWorthWidget'
-import { RecentActivityWidget } from '@/dashboard/widgets/RecentActivityWidget'
-import { RunwayWidget } from '@/dashboard/widgets/RunwayWidget'
-import { SavingsRateWidget } from '@/dashboard/widgets/SavingsRateWidget'
-import { SpendingBreakdownWidget } from '@/dashboard/widgets/SpendingBreakdownWidget'
-import { SpendingComparisonWidget } from '@/dashboard/widgets/SpendingComparisonWidget'
-import { TopBudgetsWidget } from '@/dashboard/widgets/TopBudgetsWidget'
+import { CreditWidget } from '@/dashboard/widgets/credit/CreditWidget'
+import { NetWorthWidget } from '@/dashboard/widgets/net-worth/NetWorthWidget'
+import { RecentActivityWidget } from '@/dashboard/widgets/recent-activity/RecentActivityWidget'
+import { RunwayWidget } from '@/dashboard/widgets/runway/RunwayWidget'
+import { SavingsRateWidget } from '@/dashboard/widgets/savings-rate/SavingsRateWidget'
+import { SpendingBreakdownWidget } from '@/dashboard/widgets/spending-breakdown/SpendingBreakdownWidget'
+import { SpendingComparisonWidget } from '@/dashboard/widgets/spending-comparison/SpendingComparisonWidget'
+import { TopBudgetsWidget } from '@/dashboard/widgets/top-budgets/TopBudgetsWidget'
 import { useDashboardGreeting } from '@/dashboard/hooks/useDashboardGreeting'
 
 /**
