+++ frontend/src/dashboard/widgets/net-worth/NetWorthChart.tsx
