+++ frontend/src/dashboard/widgets/net-worth/NetWorthHeader.tsx
