+++ frontend/src/dashboard/widgets/net-worth/NetWorthMetric.tsx
