+++ frontend/src/dashboard/widgets/runway/RunwaySegmentsBar.tsx
