+++ frontend/src/dashboard/widgets/savings-rate/SavingsRateChart.tsx
