+++ frontend/src/dashboard/widgets/savings-rate/SavingsRateWidget.tsx
@@ -2,8 +2,8 @@ import { useMemo, useState } from 'react'
 import { useDashboardSavingsRate } from '@/api/dashboard'
 import { useLoadingSnapshot } from '@/components/useLoadingSnapshot'
 import { DashboardWidgetLoadingBody } from '@/dashboard/components/DashboardWidgetLoadingBody'
-import { SavingsRateChart } from '@/dashboard/components/SavingsRateChart'
-import { SavingsRateHeader } from '@/dashboard/components/SavingsRateHeader'
+import { SavingsRateChart } from './SavingsRateChart'
+import { SavingsRateHeader } from './SavingsRateHeader'
 import { getSavingsRateChartData } from '@/dashboard/utils/getSavingsRateChartData'
 import { getSavingsRateSeries } from '@/dashboard/utils/getSavingsRateSeries'
