+++ frontend/src/dashboard/widgets/spending-breakdown/SpendingBreakdownCrossoverBadge.tsx
