+++ frontend/src/dashboard/widgets/spending-comparison/SpendingComparisonHeader.tsx
