+++ frontend/src/dashboard/widgets/spending-comparison/SpendingComparisonLegend.tsx
