+++ frontend/src/dashboard/widgets/spending-comparison/SpendingComparisonMetric.tsx
