+++ frontend/src/dashboard/widgets/spending-comparison/SpendingComparisonTooltipContent.tsx
