+++ frontend/src/dashboard/widgets/top-budgets/TopBudgetsWidget.tsx
@@ -2,8 +2,8 @@ import { useMemo } from 'react'
 import { useLatestBudgetUtilizations } from '@/api/budgets'
 import { useLoadingSnapshot } from '@/components/useLoadingSnapshot'
 import { DashboardWidgetLoadingBody } from '@/dashboard/components/DashboardWidgetLoadingBody'
-import { TopBudgetsHeader } from '@/dashboard/components/TopBudgetsHeader'
-import { TopBudgetsList } from '@/dashboard/components/TopBudgetsList'
+import { TopBudgetsHeader } from './TopBudgetsHeader'
+import { TopBudgetsList } from './TopBudgetsList'
 import { combineFxStatuses } from '@/dashboard/utils/fxStatus'
 import { getTopBudgets } from '@/dashboard/utils/getTopBudgets'
